+++ bin/download_emails.php
@@ -25,9 +25,6 @@
 // | 51 Franklin Street, Suite 330                                        |
 // | Boston, MA 02110-1301, USA.                                          |
 // +----------------------------------------------------------------------+
-// | Authors: João Prado Maia <jpm@mysql.com>                             |
-// | Authors: Elan Ruusamäe <glen@delfi.ee>                               |
-// +----------------------------------------------------------------------+
 
 // Set to almost 2Gigs, as 2048M overflows:
 // script tried to disable memory_limit by setting it to a negative value -2147483648 bytes which is not allowed