+++ bin/irc-bot.php
@@ -1,12 +1,12 @@
+#!/usr/bin/php
 <?php
-
 /* vim: set expandtab tabstop=4 shiftwidth=4 encoding=utf-8: */
 // +----------------------------------------------------------------------+
 // | Eventum - Issue Tracking System                                      |
 // +----------------------------------------------------------------------+
 // | Copyright (c) 2003 - 2008 MySQL AB                                   |
 // | Copyright (c) 2008 - 2010 Sun Microsystem Inc.                       |
-// | Copyright (c) 2011 - 2013 Eventum Team.                              |
+// | Copyright (c) 2011 - 2015 Eventum Team.                              |
 // |                                                                      |
 // | This program is free software; you can redistribute it and/or modify |
 // | it under the terms of the GNU General Public License as published by |
@@ -22,38 +22,46 @@
 // | along with this program; if not, write to:                           |
 // |                                                                      |
 // | Free Software Foundation, Inc.                                       |
-// | 51 Franklin Street, Suite 330                                          |
+// | 51 Franklin Street, Suite 330                                        |
 // | Boston, MA 02110-1301, USA.                                          |
 // +----------------------------------------------------------------------+
-// | Authors: Bryan Alsdorf <bryan@mysql.com>                             |
-// +----------------------------------------------------------------------+
 
-// This is an example config file for the IRC bot. This file should be
-// placed in the config/ directory.
+ini_set('memory_limit', '1024M');
 
+require_once __DIR__ . '/../init.php';
 
-// IRC server address
-$irc_server_hostname = 'localhost';
-$irc_server_port = 6667;
+// if requested, clear the lock
+if (in_array('--fix-lock', $argv)) {
+    Lock::release('irc_bot');
+    echo "The lock file was removed successfully.\n";
+    exit;
+}
 
-// the following is the list of IRC channels that the bot should connect to,
-// and the associated project name
-//      Project Name -> IRC Channel(s),
-//      Second Project' => array('#issues_2', '#byrocrate'),
-// If you want to use IRC message categories (only applies if you have a custom workflow backend)
-// the list of channels should be an associated array with the channel for the key and an array of categories
-// for the value:
-//      Project => array(
-//          '#issues_2' =>  array(APP_EVENTUM_IRC_CATEGORY_DEFAULT, 'other')
-//      )
-$irc_channels = array(
-    'Default Project' => '#issues',
-);
+if (in_array('--check-process', $argv)) {
+    $check = true;
+} else {
+    $check = false;
+}
 
-$nickname = 'EventumBOT';
-$realname = 'Eventum Issue Tracking System';
+// NB: must require this in global context
+// otherise $SMARTIRC_nreplycodes from defines.php is not initialized
+require_once 'Net/SmartIRC/defines.php';
 
-// do you need a username/password to connect to this server? if
-// so, fill in the next two variables
-$username = '';
-$password = '';
+try {
+    $bot = new Eventum_Bot();
+} catch (InvalidArgumentException $e) {
+    error_log($e->getMessage());
+    error_log("Please see config/irc_config.dist.php for sample config.");
+    exit(1);
+}
+
+// acquire a lock to prevent multiple scripts from
+// running at the same time
+if (!$bot->lock($check)) {
+    error_log('Error: Another instance of the script is still running.');
+    error_log("If this is not accurate, you may fix it by running this script with '--fix-lock' as the only parameter.");
+    exit(1);
+}
+
+$bot->run();
+$bot->unlock();