+++ htdocs/notification.php
@@ -25,9 +25,6 @@
 // | 51 Franklin Street, Suite 330                                        |
 // | Boston, MA 02110-1301, USA.                                          |
 // +----------------------------------------------------------------------+
-// | Authors: João Prado Maia <jpm@mysql.com>                             |
-// | Authors: Elan Ruusamäe <glen@delfi.ee>                               |
-// +----------------------------------------------------------------------+
 
 require_once __DIR__ . '/../init.php';
 
@@ -36,35 +33,40 @@ $tpl->setTemplate('notification.tpl.html');
 
 Auth::checkAuthentication('index.php?err=5', true);
 
+$issue_id = isset($_POST['issue_id']) ? (int)$_POST['issue_id'] : (int)$_GET['iss_id'];
 $usr_id = Auth::getUserID();
-$prj_id = Auth::getCurrentProject();
-$issue_id = isset($_POST['issue_id']) ? (int) $_POST['issue_id'] : (int) $_GET['iss_id'];
-$tpl->assign('issue_id', $issue_id);
 
-// assign default to avoid warnings
-$tpl->assign('info', null);
-
-if (!Access::canViewNotificationList($issue_id, Auth::getUserID())) {
+if (!Access::canViewNotificationList($issue_id, $usr_id)) {
     $tpl->setTemplate('permission_denied.tpl.html');
     $tpl->displayTemplate();
     exit;
 }
 
-// format default actions properly
-$default = Notification::getDefaultActions();
-// first setup defaults
-$res = array(
-    'updated' => 0,
-    'closed' => 0,
-    'files' => 0,
-    'emails' => 0,
-);
-foreach ($default as $action) {
-    $res[$action] = 1;
-}
-$tpl->assign('default_actions', $res);
+$sub_id = isset($_GET['id']) ? (int)$_GET['id'] : null;
+$prj_id = Auth::getCurrentProject();
+$default_actions = Notification::getDefaultActions();
 
-$cat = isset($_POST['cat']) ? (string) $_POST['cat'] : (isset($_GET['cat']) ? (string) $_GET['cat'] : null);
+if ($sub_id) {
+    $info = Notification::getDetails($sub_id);
+} else {
+    $info = array(
+        'updated' => 0,
+        'closed' => 0,
+        'files' => 0,
+        'emails' => 0,
+    );
+    foreach ($default_actions as $action) {
+        $res[$action] = 1;
+    }
+}
+
+$tpl->assign(array(
+    'issue_id' => $issue_id,
+    'default_actions' => $default_actions,
+    'info' => $info,
+));
+
+$cat = isset($_POST['cat']) ? (string)$_POST['cat'] : (isset($_GET['cat']) ? (string)$_GET['cat'] : null);
 
 if ($cat == 'insert') {
     $res = Notification::subscribeEmail($usr_id, $issue_id, $_POST['email'], $_POST['actions']);
@@ -82,8 +84,6 @@ if ($cat == 'insert') {
     }
     Auth::redirect(APP_RELATIVE_URL . 'notification.php?iss_id=' . $issue_id);
 } elseif ($cat == 'edit') {
-    $res = Notification::getDetails($_GET['id']);
-    $tpl->assign('info', $res);
 } elseif ($cat == 'delete') {
     $res = Notification::remove($_POST['items']);
     if ($res == 1) {
@@ -92,8 +92,10 @@ if ($cat == 'insert') {
 }
 
 $tpl->assign('list', Notification::getSubscriberListing($issue_id));
-$t = Project::getAddressBook($prj_id, $issue_id);
-$tpl->assign('assoc_users', $t);
+/*
+// the autocomplete is removed, no need to fetch the data
+$tpl->assign('assoc_users', Project::getAddressBook($prj_id, $issue_id));
 $tpl->assign('allowed_emails', Project::getAddressBookEmails($prj_id, $issue_id));
+*/
 
 $tpl->displayTemplate();