+++ init.php
@@ -24,9 +24,6 @@
 // | 51 Franklin Street, Suite 330                                        |
 // | Boston, MA 02110-1301, USA.                                          |
 // +----------------------------------------------------------------------+
-// | Authors: Bryan Alsdorf <bryan@mysql.com>                             |
-// | Authors: Elan Ruusamäe <glen@delfi.ee>                               |
-// +----------------------------------------------------------------------+
 
 if (!file_exists(__DIR__ . '/config/config.php') || !filesize(__DIR__ . '/config/config.php')) {
     header('Location: setup/');
@@ -40,12 +37,12 @@ ini_set('memory_limit', '512M');
 ini_set('session.cache_limiter', 'nocache');
 
 define('APP_URL', 'https://github.com/eventum/eventum');
-define('APP_VERSION', '3.0.4-dev');
+define('APP_VERSION', '3.0.5-dev');
 
 // define base path
 define('APP_PATH', __DIR__);
 
-$define = function($name, $value) {
+$define = function ($name, $value) {
     if (defined($name)) {
         return;
     }
@@ -69,16 +66,18 @@ require_once APP_CONFIG_PATH . '/config.php';
 $define('APP_LOCAL_PATH', APP_CONFIG_PATH);
 $define('APP_COOKIE', 'eventum');
 
+// /var path for writable data
+$define('APP_VAR_PATH', APP_PATH . '/var');
+
 // define other paths
 $define('APP_SETUP_FILE', APP_CONFIG_PATH . '/setup.php');
-$define('APP_TPL_PATH', APP_PATH . '/templates');
-$define('APP_TPL_COMPILE_PATH', APP_PATH . '/templates_c');
 $define('APP_INC_PATH', APP_PATH . '/lib/eventum');
-$define('APP_LOCKS_PATH', APP_PATH . '/locks');
-$define('APP_LOG_PATH', APP_PATH . '/logs');
+$define('APP_TPL_PATH', APP_PATH . '/templates');
+$define('APP_TPL_COMPILE_PATH', APP_VAR_PATH . '/cache');
+$define('APP_LOCKS_PATH', APP_VAR_PATH . '/lock');
+$define('APP_LOG_PATH', APP_VAR_PATH . '/log');
 $define('APP_ERROR_LOG', APP_LOG_PATH . '/errors.log');
 $define('APP_CLI_LOG', APP_LOG_PATH . '/cli.log');
-$define('APP_IRC_LOG', APP_LOG_PATH . '/irc_bot.log');
 $define('APP_LOGIN_LOG', APP_LOG_PATH . '/login_attempts.log');
 
 // define the user_id of system user
@@ -169,3 +168,8 @@ if (!defined('APP_DEFAULT_DB') || !defined('APP_TABLE_PREFIX')) {
     unset($dbconfig);
 }
 */
+
+// setup debugbar, if it can be autoloaded
+if (class_exists('DebugBar\StandardDebugBar')) {
+    $debugbar = new DebugBar\StandardDebugBar();
+}