+++ lib/eventum/class.custom_field.php
@@ -25,9 +25,6 @@
 // | 51 Franklin Street, Suite 330                                        |
 // | Boston, MA 02110-1301, USA.                                          |
 // +----------------------------------------------------------------------+
-// | Authors: João Prado Maia <jpm@mysql.com>                             |
-// | Authors: Elan Ruusamäe <glen@delfi.ee>                               |
-// +----------------------------------------------------------------------+
 
 
 /**
@@ -278,7 +275,7 @@ class Custom_Field
                 }
             }
 
-            Workflow::handleCustomFieldsUpdated($prj_id, $issue_id, $old_values, self::getValuesByIssue($prj_id, $issue_id));
+            Workflow::handleCustomFieldsUpdated($prj_id, $issue_id, $old_values, self::getValuesByIssue($prj_id, $issue_id), $updated_fields);
             Issue::markAsUpdated($issue_id);
             // need to save a history entry for this
 
@@ -890,7 +887,7 @@ class Custom_Field
                     ?, ?
                  )';
         try {
-            $res = DB_Helper::getInstance()->query($stmt, array($prj_id, $fld_id));
+            DB_Helper::getInstance()->query($stmt, array($prj_id, $fld_id));
         } catch (DbException $e) {
             return false;
         }