+++ tests/LinkFilterTest.php
@@ -1,12 +1,10 @@
 <?php
 
-class LinkFilterTest extends PHPUnit_Framework_TestCase
+class LinkFilterTest extends TestCase
 {
     public function setUp()
     {
-        if (getenv('TRAVIS') || getenv('JENKINS_HOME')) {
-            $this->markTestSkipped("No DB tests in Travis/Jenkins");
-        }
+        $this->skipCi("No DB tests in Travis/Jenkins");
     }
 
     /**