+++ tests/MailStorageTest.php
@@ -1,8 +1,13 @@
 <?php
 use Zend\Mail;
 
-class MailStorageTest extends PHPUnit_Framework_TestCase
+class MailStorageTest extends TestCase
 {
+    public static function setUpBeforeClass()
+    {
+        self::skipCi("Not ready yet");
+    }
+
     /** @var array */
     private $account;